requirements.txt
asn1crypto==1.5.1
cffi==1.15.0
cfgv==3.3.1
click==8.1.3
cryptography==37.0.3
distlib==0.3.4
filelock==3.7.1
Flask==2.1.2
greenlet==1.1.2
identify==2.5.1
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
nodeenv==1.6.0
pg8000==1.29.1
platformdirs==2.5.2
pre-commit==2.19.0
psycopg2-binary==2.9.3
pycparser==2.21
PyYAML==6.0
scramp==1.4.1
six==1.16.0
SQLAlchemy==1.4.37
toml==0.10.2
virtualenv==20.14.1
Werkzeug==2.1.2
zipp==3.8.0
